+++ src/Konsole.Samples/Samples/ProgressBarSamples.cs
@@ -10,35 +10,29 @@ namespace Konsole.Samples
 {
     public class ProgressBarSamples
     {
-        static void Samples()
+
+        public static void Run()
         {
+            void Pause() => Console.ReadKey(true);
 
             Console.WriteLine("progress bar in a floating inline window");
             // --------------------------------------------------------
-
-            var w = new Window(80, 10, White, DarkCyan);
+            Pause();
+            var w = new Window(80, 10);
             var pb1 = new ProgressBar(w, 100);
             var pb2 = new ProgressBar(w, 100);
             pb1.Refresh(25, "hello world! 1");
             pb2.Refresh(25, "hello world! 2");
-            Console.ReadKey(true);
-
-            Console.WriteLine("progress bar in a floating inline window with title box");
-            // --------------------------------------------------------------------------
-
-            var left = w.SplitLeft("left");
-            var right = w.SplitRight("right");
-            var pb = new ProgressBar(left, 100);
-            pb.Refresh(25, "hello world!");
-            Console.ReadKey(true);
+            
+            Pause();
 
             Console.WriteLine("progress bar in a floating inline window + split window LeftRight");
             // -----------------------------------------------------------------------------------
 
             w = new Window(80, 10, White, DarkBlue);
-            left = w.SplitLeft("left");
-            right = w.SplitRight("right");
-            pb = new ProgressBar(left, 100);
+            var left = w.SplitLeft("left");
+            var right = w.SplitRight("right");
+            var pb = new ProgressBar(left, 100);
             pb.Refresh(25, "hello world!");
             Console.ReadKey(true);